requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
dateparser==1.2.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
multidict==6.0.5
numpy==1.26.4
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.1
TA-Lib==0.4.28
typing_extensions==4.10.0
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
websocket-client==1.7.0
websockets==12.0
yarl==1.9.4
